\name{nc_version}
\alias{nc_version}
\title{Report version of ncdf4 library}
\description{
 Returns a string that is the version number of the ncdf4 package.
}
\usage{
 nc_version()
}
\value{
 A string (not float) that is the version number of the ncdf4 package.
}
\references{
 http://dwpierce.com/software
}
\details{
 Note that the returned value it is a string, not a floating point number.
}
\author{David W. Pierce \email{dpierce@ucsd.edu}}
\keyword{utilities}
